Crate mlx_rs

Modules

error

fft

Fast Fourier Transform (FFT) and its inverse (IFFT) for one, two, and N dimensions.

linalg

ops

prelude

random

transforms

utils

Macros

array

A helper macro to create an array with up to 3 dimensions.

assert_array_eq

Structs

Array

Device

Representation of a Device in MLX.

DtypeIter

An iterator over the variants of Dtype

Stream

A stream of evaluation attached to a particular device.

StreamOrDevice

Parameter type for all MLX operations.

Enums

DeviceType

Type of device.

Dtype

Array element type

Traits

ArrayElement

A marker trait for array elements.

FromNested

A helper trait to construct Array from nested arrays or slices.

FromScalar

A helper trait to construct Array from scalar values.

Functions

stop_gradient

Stop gradients from being computed.

stop_gradient_device

Stop gradients from being computed.

Type Aliases

complex64