Function mlx_rs::fft::rfftn_device

pub fn rfftn_device<'a>(
    a: &Array,
    s: impl IntoOption<&'a [i32]>,
    axes: impl IntoOption<&'a [i32]>,
    stream: impl AsRef<Stream>
) -> Result<Array, Exception>

n-dimensional real discrete Fourier Transform.

The output has the same shape as the input except along the dimensions in axes in which case it has sizes from s. The last axis in axes is treated as the real axis and will have size s[s.len()-1] // 2 + 1.

Params

a: The input array. If the array is complex it will be silently cast to a real type.

s: Sizes of the transformed axes. The corresponding axes in the input are truncated or padded with zeros to match s. The default value is the sizes of a along axes.

axes: Axes along which to perform the FFT. The default is None in which case the FFT is over the last len(s) axes or all axes if s is also None.