Module mlx_rs::linalg

Enums

Ord

Functions

cholesky

Compute the Cholesky decomposition of a real symmetric positive semi-definite matrix.

inv

Compute the inverse of a square matrix. Returns an error if the input is not valid.

inv_device

Compute the inverse of a square matrix. Returns an error if the input is not valid.

norm

Matrix or vector norm.

norm_device

Matrix or vector norm.

norm_ord

Matrix or vector norm.

norm_ord_device

Matrix or vector norm.

norm_p

norm_p_device

qr

The QR factorization of the input matrix. Returns an error if the input is not valid.

qr_device

The QR factorization of the input matrix. Returns an error if the input is not valid.

svd

The Singular Value Decomposition (SVD) of the input matrix. Returns an error if the input is not valid.

svd_device

The Singular Value Decomposition (SVD) of the input matrix. Returns an error if the input is not valid.