Function mlx_rs::linalg::cholesky

pub fn cholesky(
    a: &Array,
    upper: Option<bool>,
    stream: impl AsRef<Stream>,
) -> Result<Array, Exception>

Compute the Cholesky decomposition of a real symmetric positive semi-definite matrix.

This function supports arrays with at least 2 dimensions. When the input has more than two dimensions, the Cholesky decomposition is computed for each matrix in the last two dimensions of a.

If the input matrix is not symmetric positive semi-definite, behaviour is undefined.

Params

a: input array

upper: If true, return the upper triangular Cholesky factor. If false, return the lower triangular Cholesky factor. Default: false.