Module mlx_rs::ops

Modules

indexing

Indexing Arrays

Enums

PadWidth

TensorDotDims

Traits

ClipBound

A custom trait for the bound of the clip operation.

Functions

abs

Element-wise absolute value.

abs_device

Element-wise absolute value.

acos

Element-wise inverse cosine.

acos_device

Element-wise inverse cosine.

acosh

Element-wise inverse hyperbolic cosine.

acosh_device

Element-wise inverse hyperbolic cosine.

add

See Array::add.

add_device

See Array::add.

addmm

Matrix multiplication with addition and optional scaling.

addmm_device

Matrix multiplication with addition and optional scaling.

all

See Array::all

all_close

See Array::all_close

all_close_device

See Array::all_close

all_device

See Array::all

any

See Array::any

any_device

See Array::any

arange

See Array::arange

arange_device

See Array::arange

argpartition

Returns the indices that partition the array. Returns an error if the arguments are invalid.

argpartition_all

Returns the indices that partition the flattened array. Returns an error if the arguments are invalid.

argpartition_all_device

Returns the indices that partition the flattened array. Returns an error if the arguments are invalid.

argpartition_device

Returns the indices that partition the array. Returns an error if the arguments are invalid.

argsort

Returns the indices that sort the array. Returns an error if the arguments are invalid.

argsort_all

Returns the indices that sort the flattened array. Returns an error if the arguments are invalid.

argsort_all_device

Returns the indices that sort the flattened array. Returns an error if the arguments are invalid.

argsort_device

Returns the indices that sort the array. Returns an error if the arguments are invalid.

array_eq

See Array::array_eq

array_eq_device

See Array::array_eq

as_strided

Create a view into the array with the given shape and strides.

as_strided_device

Create a view into the array with the given shape and strides.

asin

Element-wise inverse sine.

asin_device

Element-wise inverse sine.

asinh

Element-wise inverse hyperbolic sine.

asinh_device

Element-wise inverse hyperbolic sine.

at_least_1d

Convert array to have at least one dimension.

at_least_1d_device

Convert array to have at least one dimension.

at_least_2d

Convert array to have at least two dimensions.

at_least_2d_device

Convert array to have at least two dimensions.

at_least_3d

Convert array to have at least three dimensions.

at_least_3d_device

Convert array to have at least three dimensions.

atan

Element-wise inverse tangent.

atan_device

Element-wise inverse tangent.

atanh

Element-wise inverse hyperbolic tangent.

atanh_device

Element-wise inverse hyperbolic tangent.

block_masked_mm

Matrix multiplication with block masking.

block_masked_mm_device

Matrix multiplication with block masking.

broadcast_arrays

Broadcast a vector of arrays against one another. Returns an error if the shapes are broadcastable.

broadcast_arrays_device

Broadcast a vector of arrays against one another. Returns an error if the shapes are broadcastable.

broadcast_to

Broadcast an array to the given shape. Returns an error if the shapes are not broadcastable.

broadcast_to_device

Broadcast an array to the given shape. Returns an error if the shapes are not broadcastable.

ceil

Element-wise ceiling.

ceil_device

Element-wise ceiling.

clip

Clip the values of the array between the given minimum and maximum.

clip_device

Clip the values of the array between the given minimum and maximum.

concatenate

Concatenate the arrays along the given axis. Returns an error if the shapes are invalid.

concatenate_device

Concatenate the arrays along the given axis. Returns an error if the shapes are invalid.

conv1d

1D convolution over an input with several channels returning an error if the inputs are invalid.

conv1d_device

1D convolution over an input with several channels returning an error if the inputs are invalid.

conv2d

2D convolution over an input with several channels returning an error if the inputs are invalid.

conv2d_device

2D convolution over an input with several channels returning an error if the inputs are invalid.

conv_general

General convolution over an input with several channels returning an error if the inputs are invalid.

conv_general_device

General convolution over an input with several channels returning an error if the inputs are invalid.

cos

Element-wise cosine.

cos_device

Element-wise cosine.

cosh

Element-wise hyperbolic cosine.

cosh_device

Element-wise hyperbolic cosine.

cummax

See Array::cummax

cummax_device

See Array::cummax

cummin

See Array::cummin

cummin_device

See Array::cummin

cumprod

See Array::cumprod

cumprod_device

See Array::cumprod

cumsum

See Array::cumsum

cumsum_device

See Array::cumsum

degrees

Convert angles from radians to degrees.

degrees_device

Convert angles from radians to degrees.

dequantize

Dequantize the matrix w using the provided scales and biases and the group_size and bits configuration.

dequantize_device

Dequantize the matrix w using the provided scales and biases and the group_size and bits configuration.

diag

See Array::diag

diag_device

See Array::diag

diagonal

See Array::diagonal

diagonal_device

See Array::diagonal

divide

See Array::divide.

divide_device

See Array::divide.

divmod

Element-wise quotient and remainder.

divmod_device

Element-wise quotient and remainder.

eq

See Array::eq

eq_device

See Array::eq

erf

Element-wise error function.

erf_device

Element-wise error function.

erfinv

Element-wise inverse error function.

erfinv_device

Element-wise inverse error function.

exp

See Array::exp.

exp_device

See Array::exp.

expand_dims

Add a size one dimension at the given axis, returns an error if the axes are invalid.

expand_dims_device

Add a size one dimension at the given axis, returns an error if the axes are invalid.

expm1

Element-wise exponential minus 1.

expm1_device

Element-wise exponential minus 1.

eye

See Array::eye

eye_device

See Array::eye

flatten

Flatten an array. Returns an error if the axes are invalid.

flatten_device

Flatten an array. Returns an error if the axes are invalid.

floor

See Array::floor.

floor_device

See Array::floor.

floor_divide

See Array::floor_divide.

floor_divide_device

See Array::floor_divide.

full

See Array::full

full_device

See Array::full

ge

See Array::ge

ge_device

See Array::ge

gt

See Array::gt

gt_device

See Array::gt

identity

See Array::identity

identity_device

See Array::identity

inner

Ordinary inner product of vectors for 1-D arrays, in higher dimensions a sum product over the last axes.

inner_device

Ordinary inner product of vectors for 1-D arrays, in higher dimensions a sum product over the last axes.

is_close

See Array::is_close

is_close_device

See Array::is_close

is_inf

Return a boolean array indicating which elements are +/- inifnity.

is_inf_device

Return a boolean array indicating which elements are +/- inifnity.

is_nan

Return a boolean array indicating which elements are NaN.

is_nan_device

Return a boolean array indicating which elements are NaN.

is_neg_inf

Return a boolean array indicating which elements are negative infinity.

is_neg_inf_device

Return a boolean array indicating which elements are negative infinity.

is_pos_inf

Return a boolean array indicating which elements are positive infinity.

is_pos_inf_device

Return a boolean array indicating which elements are positive infinity.

le

See Array::le

le_device

See Array::le

linspace

See Array::linspace

linspace_device

See Array::linspace

log

See Array::log.

log1p

See Array::log1p.

log1p_device

See Array::log1p.

log2

See Array::log2.

log2_device

See Array::log2.

log10

See Array::log10.

log10_device

See Array::log10.

log_add_exp

Element-wise log-add-exp.

log_add_exp_device

Element-wise log-add-exp.

log_device

See Array::log.

log_sum_exp

See Array::log_sum_exp

log_sum_exp_device

See Array::log_sum_exp

logical_and

See Array::logical_and

logical_and_device

See Array::logical_and

logical_not

See Array::logical_not

logical_not_device

See Array::logical_not

logical_or

See Array::logical_or

logical_or_device

See Array::logical_or

lt

See Array::lt

lt_device

See Array::lt

matmul

See Array::matmul.

matmul_device

See Array::matmul.

max

See Array::max

max_device

See Array::max

maximum

Element-wise maximum.

maximum_device

Element-wise maximum.

mean

See Array::mean

mean_device

See Array::mean

min

See Array::min

min_device

See Array::min

minimum

Element-wise minimum.

minimum_device

Element-wise minimum.

move_axis

Move an axis to a new position. Returns an error if the axes are invalid.

move_axis_device

Move an axis to a new position. Returns an error if the axes are invalid.

multiply

See Array::multiply.

multiply_device

See Array::multiply.

ne

See Array::ne

ne_device

See Array::ne

negative

See Array::negative.

negative_device

See Array::negative.

ones

See Array::ones

ones_device

See Array::ones

outer

Compute the outer product of two 1-D arrays, if the array’s passed are not 1-D a flatten op will be run beforehand.

outer_device

Compute the outer product of two 1-D arrays, if the array’s passed are not 1-D a flatten op will be run beforehand.

pad

Pad an array with a constant value. Returns an error if the width is invalid.

pad_device

Pad an array with a constant value. Returns an error if the width is invalid.

partition

Returns a partitioned copy of the array such that the smaller kth elements are first. Returns an error if the arguments are invalid.

partition_all

Returns a partitioned copy of the flattened array such that the smaller kth elements are first. Returns an error if the arguments are invalid.

partition_all_device

Returns a partitioned copy of the flattened array such that the smaller kth elements are first. Returns an error if the arguments are invalid.

partition_device

Returns a partitioned copy of the array such that the smaller kth elements are first. Returns an error if the arguments are invalid.

power

See Array::power.

power_device

See Array::power.

prod

See Array::prod

prod_device

See Array::prod

quantize

Quantize the matrix w using bits bits per element.

quantize_device

Quantize the matrix w using bits bits per element.

quantized_matmul

Perform the matrix multiplication with the quantized matrix w. The quantization uses one floating point scale and bias per group_size of elements. Each element in w takes bits bits and is packed in an unsigned 32 bit integer.

quantized_matmul_device

Perform the matrix multiplication with the quantized matrix w. The quantization uses one floating point scale and bias per group_size of elements. Each element in w takes bits bits and is packed in an unsigned 32 bit integer.

radians

Convert angles from degrees to radians.

radians_device

Convert angles from degrees to radians.

reciprocal

See Array::reciprocal.

reciprocal_device

See Array::reciprocal.

remainder

See Array::remainder.

remainder_device

See Array::remainder.

repeat

See Array::repeat

repeat_all

See Array::repeat_all

repeat_all_device

See Array::repeat_all

repeat_device

See Array::repeat

reshape

Reshape an array while preserving the size. Returns an error if the new shape is invalid.

reshape_device

Reshape an array while preserving the size. Returns an error if the new shape is invalid.

round

See Array::round.

round_device

See Array::round.

rsqrt

See Array::rsqrt.

rsqrt_device

See Array::rsqrt.

sigmoid

Element-wise logistic sigmoid.

sigmoid_device

Element-wise logistic sigmoid.

sign

Element-wise sign.

sign_device

Element-wise sign.

sin

See Array::sin.

sin_device

See Array::sin.

sinh

Element-wise hyperbolic sine.

sinh_device

Element-wise hyperbolic sine.

softmax

Perform the softmax along the given axis.

softmax_device

Perform the softmax along the given axis.

sort

Returns a sorted copy of the array. Returns an error if the arguments are invalid.

sort_all

Returns a sorted copy of the flattened array. Returns an error if the arguments are invalid.

sort_all_device

Returns a sorted copy of the flattened array. Returns an error if the arguments are invalid.

sort_device

Returns a sorted copy of the array. Returns an error if the arguments are invalid.

split

Split an array along a given axis. Returns an error if the indices are invalid.

split_device

Split an array along a given axis. Returns an error if the indices are invalid.

split_equal

Split an array into equal parts along a given axis. Returns an error if the array cannot be split into equal parts.

split_equal_device

Split an array into equal parts along a given axis. Returns an error if the array cannot be split into equal parts.

sqrt

See Array::sqrt.

sqrt_device

See Array::sqrt.

square

See Array::square.

square_device

See Array::square.

squeeze

Remove length one axes from an array. Returns an error if the axes are invalid.

squeeze_device

Remove length one axes from an array. Returns an error if the axes are invalid.

stack

Stacks the arrays along a new axis. Returns an error if the arguments are invalid.

stack_all

Stacks the arrays along a new axis. Returns an error if the arrays have different shapes.

stack_all_device

Stacks the arrays along a new axis. Returns an error if the arrays have different shapes.

stack_device

Stacks the arrays along a new axis. Returns an error if the arguments are invalid.

std

Compute the standard deviation(s) over the given axes.

std_device

Compute the standard deviation(s) over the given axes.

subtract

See Array::subtract.

subtract_device

See Array::subtract.

sum

See Array::sum

sum_device

See Array::sum

swap_axes

Swap two axes of an array. Returns an error if the axes are invalid.

swap_axes_device

Swap two axes of an array. Returns an error if the axes are invalid.

tan

See [Array::tan].

tan_device

See [Array::tan].

tanh

Element-wise hyperbolic tangent.

tanh_device

Element-wise hyperbolic tangent.

tensordot

Compute the tensor dot product along the specified axes.

tensordot_device

Compute the tensor dot product along the specified axes.

tile

Construct an array by repeating a the number of times given by reps.

tile_device

Construct an array by repeating a the number of times given by reps.

transpose

Transpose the dimensions of the array. Returns an error if the axes are invalid.

transpose_device

Transpose the dimensions of the array. Returns an error if the axes are invalid.

tri

See Array::tri

tri_device

See Array::tri

variance

See Array::variance

variance_device

See Array::variance

where

Select from a or b according to condition returning an error if the arrays are not broadcastable.

where_device

Select from a or b according to condition returning an error if the arrays are not broadcastable.

which

Alias for [r#where]

which_device

Alias for [r#where]

zeros

See Array::zeros

zeros_device

See Array::zeros