Function mlx_rs::ops::where

pub fn where<'a, 'b>(
    condition: &Array,
    a: impl ScalarOrArray<'a>,
    b: impl ScalarOrArray<'b>,
) -> Result<Array, Exception>

Select from a or b according to condition returning an error if the arrays are not broadcastable.

The condition and input arrays must be the same shape or broadcasting with each another.

Params

condition: condition array

a: input selected from where condition is non-zero or true

b: input selected from where condition is zero or false