Trait mlx_rs::FromNested

pub trait FromNested<T> {
    // Required method
    fn from_nested(data: T) -> Array;
}

A helper trait to construct Array from nested arrays or slices.

Given that this is not intended for use other than the macro [array!], we added this trait instead of directly implementing From for Array to avoid conflicts with other From implementations.

Beware that this is subject to change in the future should we find a better way to implement the macro without creating conflicts.

Required Methods

fn from_nested(data: T) -> Array

Object Safety

This trait is not object safe.