Module mlx_rs::transforms

Modules

compile

Traits

Grad

Functions

async_eval

Asynchronously evaluate an iterator of Arrays.

eval

Evaluate an iterator of Arrays.

grad

Returns a function which computes the gradient of f.

jvp

Compute the Jacobian-vector product.

value_and_grad

Returns a function which computes the value and gradient of f.

vjp

Compute the vector-Jacobian product.